+++ src/main.rs
@@ -1,4 +1,3 @@
-//!
 //! A cargo subcommand for checking the latest version on crates.io of a particular dependency
 //!
 //! ## About
@@ -63,9 +62,11 @@
 //!     -v, --verbose           Print verbose output
 //!
 //! OPTIONS:
-//!     -d, --depth <DEPTH>       How deep in the dependency chain to search
-//!                               (Defaults to all dependencies when omitted)
-//!     -p, --package <PKG>...    Package to inspect for updates
+//!     -d, --depth <NUM>             How deep in the dependency chain to search (Defaults to all dependencies when omitted)
+//!         --exit-code <NUM>         The exit code to return on new versions found [default: 0]
+//!     -l, --lockfile-path <PATH>    An absolute path to the Cargo.lock to use (Defaults to Cargo.lock in project root)
+//!     -m, --manifest-path <PATH>    An absolute path to the Cargo.toml to use (Defaults to Cargo.toml in project root)
+//!     -p, --package <PKG>...        Package to inspect for updates
 //! ```
 //!
 //! ## License
@@ -101,72 +102,77 @@ mod lockfile;
 mod deps;
 mod error;
 mod fmt;
+mod util;
 
 use std::io::{Write, stdout};
+use std::path::Path;
 #[cfg(feature="debug")]
 use std::env;
 use std::process;
 
-use clap::{App, AppSettings, Arg, SubCommand};
+use clap::{App, AppSettings, Arg, SubCommand, ArgMatches};
 use tabwriter::TabWriter;
 
 use config::Config;
 use lockfile::Lockfile;
-use error::CliError;
+use error::CliResult;
 use fmt::Format;
 
-/// Convenience type to return a result or a `CliError`
-pub type CliResult<T> = Result<T, CliError>;
-
 fn main() {
-    debugln!("executing; cmd=cargo-outdated; args={:?}",
-             env::args().collect::<Vec<_>>());
+    debugln!("main:args={:?}", env::args().collect::<Vec<_>>());
     let m = App::new("cargo-outdated")
         .author("Kevin K. <kbknapp@gmail.com>")
         .about("Displays information about project dependency versions")
-        .version(&*format!("v{}", crate_version!()))
-    // We have to lie about our binary name since this will be a third party
-    // subcommand for cargo
+        .version(concat!("v", crate_version!()))
+        // We have to lie about our binary name since this will be a third party
+        // subcommand for cargo
         .bin_name("cargo")
-    // Global version uses the version we supplied (Cargo.toml) for all subcommands
-    // as well
+        // Global version uses the version we supplied (Cargo.toml) for all subcommands
+        // as well
         .settings(&[AppSettings::GlobalVersion,
                     AppSettings::SubcommandRequired])
-    // We use a subcommand because parsed after `cargo` is sent to the third party
-    // plugin
-    // which will be interpreted as a subcommand/positional arg by clap
+        // We use a subcommand because parsed after `cargo` is sent to the third party
+        // plugin
+        // which will be interpreted as a subcommand/positional arg by clap
         .subcommand(SubCommand::with_name("outdated")
             .about("Displays information about project dependency versions")
             .args_from_usage(
-                "-p, --package [PKG]...    'Package to inspect for updates'
-                 -v, --verbose             'Print verbose output'
-                 -d, --depth [NUM]       'How deep in the dependency chain to search \
+                "-p, --package [PKG]...     'Package to inspect for updates'
+                 -v, --verbose              'Print verbose output'
+                 -d, --depth [NUM]          'How deep in the dependency chain to search \
                                             (Defaults to all dependencies when omitted)'")
-            .arg(Arg::from_usage("--exit-code [NUM]     'The exit code to return on new versions found'")
-                .default_value("0"))
-    // We separate -R so we can addd a conflicting argument
-            .arg(Arg::from_usage(
-                "-R, --root-deps-only  'Only check root dependencies (Equivilant to --depth=1)'")
-                .conflicts_with("DEPTH")))
+            .args(&[
+                Arg::from_usage("--exit-code [NUM]     'The exit code to return on new versions found'")
+                    .default_value("0"),
+                Arg::from_usage(
+                    "-R, --root-deps-only  'Only check root dependencies (Equivilant to --depth=1)'")
+                    .conflicts_with("depth"),
+                Arg::from_usage("-m, --manifest-path [PATH] 'An absolute path to the Cargo.toml file to use \
+                                                             (Defaults to Cargo.toml in project root)'")
+                    .validator(is_file),
+                Arg::from_usage("-l, --lockfile-path [PATH] 'An absolute path to the Cargo.lock to use \
+                                                             (Defaults to Cargo.lock in project root)'")
+                    .validator(is_file)]))
         .get_matches();
 
     if let Some(m) = m.subcommand_matches("outdated") {
-        let cfg = Config::from_matches(m);
-        match execute(cfg) {
+        match execute(m) {
             Ok(code) => {
-                debugln!("exit_code={}", code);
+                debugln!("main:exit_code={}", code);
                 process::exit(code)
-            },
+            }
             Err(e) => e.exit(),
         }
     }
 }
 
-fn execute(cfg: Config) -> CliResult<i32> {
-    debugln!("executing; execute; cfg={:?}", cfg);
+fn execute(m: &ArgMatches) -> CliResult<i32> {
+    debugln!("execute:m={:#?}", m);
+    let cfg = try!(Config::from_matches(m));
 
-    verbose!(cfg, "Parsing {}...", Format::Warning("Cargo.lock"));
-    let mut lf = try!(Lockfile::new());
+    verbose!(cfg, "Parsing {}...", Format::Warning(cfg.lockfile.to_string_lossy()));
+
+    let mut lf = try!(Lockfile::from_config(&cfg));
     verboseln!(cfg, "{}", Format::Good("Done"));
 
     match lf.get_updates(&cfg) {
@@ -202,4 +208,12 @@ fn execute(cfg: Config) -> CliResult<i32> {
         }
         Err(e) => Err(e),
     }
+}
+
+fn is_file(s: String) -> Result<(), String> {
+    let p = Path::new(&*s);
+    if p.file_name().is_none() {
+        return Err(format!("'{}' doesn't appear to be a valid file name", &*s));
+    }
+    Ok(())
 }